backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.35.0
azure-data-tables==12.7.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.14
h11==0.16.0
idna==3.10
isodate==0.7.2
lxml==6.0.0
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
httpx==0.27.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
yarl==1.20.1
email-validator==2.1.0
PyJWT==2.8.0
schedule==1.2.0
playwright==1.48.0

# smtplib and email.mime are built-in modules used for sending emails
